fix res. of vars with env,config,workspaceFolder
If a variable had env, config, or workspaceFolder as substring, it was
being treated as an environment variable lookup on the reminder of the
string minus the character immediately following the substring. For
instance: envRoot was being looked up as "oot" in the configuarion
environment block.

The issue is that the regex to match the env., config., and
workspaceFolder. patterns used . instead of \.