Validate dependencies for lldb-mi
If a user does not provide a custom lldb-mi on macOS running in LLDB
mode. We check to see if the XCode/XCode CLI LLDB.framework dependencies
exist.

Also adds both LLDB-MI Package based on versionRegex

This PR adds in PlatformInfo.version and IPackage.versionRegex and
IPackage.matchVersion.

Currently only darwin PlatformInfo.version is set since only the LLDB-MI
package utilizes IPackage.versionRegex and matchVersion.

The LLDB-MI 3.8 package is used for High Sierra and lower and the newer
lldb-mi is used for Mojave and higher