Add dotConfig property for c_cpp_properties.json
Kconfig system generate a .config file with all the defines, CONFIG
options selected, to build a project. Examples of projects that use
Kconfig system is Linux Kernel and NuttX RTOS. This patch add the
`dotConfig` property that will store the path for read a .config file
and will concat the defines from file to the extension defines.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>